+++ app.go
@@ -2,14 +2,12 @@
 package warg
 
 import (
-	"errors"
 	"fmt"
 	"log"
 	"os"
 	"runtime/debug"
 	"strings"
 
-	"go.bbkane.com/warg/command"
 	"go.bbkane.com/warg/config"
 	"go.bbkane.com/warg/flag"
 	"go.bbkane.com/warg/help"
@@ -151,29 +149,6 @@ func debugBuildInfoVersion() string {
 	return info.Main.Version
 }
 
-// AddVersionCommand adds a "version" command to the root section that prints the version passed.
-// Pass an empty string to use .Main.Version from `debug.ReadBuildInfo`,
-// which returns "(devel)" when using `go run`
-func AddVersionCommand(version string) AppOpt {
-	return func(a *App) {
-		action := func(ctx command.Context) error {
-			if version != "" {
-				fmt.Println(version)
-				return nil
-			}
-			// If installed via `go install`, we'll be able to read runtime version info
-			info, ok := debug.ReadBuildInfo()
-			if !ok {
-				return errors.New("unable to read build info")
-			}
-			// when run with `go run`, this will return "(devel)"
-			fmt.Println(info.Main.Version)
-			return nil
-		}
-		section.Command("version", "Print version", action)(&a.rootSection)
-	}
-}
-
 // AddColorFlag adds a "--color" flag to the root section. By convention, this flag will be respected by the different help commands
 func AddColorFlag() AppOpt {
 	return func(a *App) {
@@ -238,7 +213,7 @@ func New(name string, rootSection section.SectionT, opts ...AppOpt) App {
 	}
 
 	if app.version == "" {
-		OverrideVersion(debugBuildInfoVersion())
+		OverrideVersion(debugBuildInfoVersion())(&app)
 	}
 
 	// validate or not and return