+++ command/command.go
@@ -71,7 +71,9 @@ func DoNothing(_ Context) error {
 	return errors.New("NOTE: replace this command.DoNothing call")
 }
 
-// PrintVersion prints the app's version
+// PrintVersion prints the app's version. Use as follows
+//
+//	section.Command("version", "Print version", command.PrintVersion),
 func PrintVersion(ctx Context) error {
 	fmt.Fprintln(ctx.Stdout, ctx.Version)
 	return nil