+++ examples/starghaze/main.go
@@ -221,6 +221,7 @@ func app() *warg.App {
 		"starghaze",
 		section.New(
 			"Save GitHub Starred Repos",
+			section.Command("version", "Print version", command.PrintVersion),
 			section.ExistingCommand(
 				"download",
 				downloadCmd,
@@ -240,7 +241,6 @@ func app() *warg.App {
 			section.Footer("Homepage: https://github.com/bbkane/starghaze"),
 		),
 		warg.AddColorFlag(),
-		warg.AddVersionCommand(version),
 		warg.SkipValidation(),
 	)
 	return &app